fix(test): update test for removing labels from execution
