+++ webserver/src/test/java/io/kestra/webserver/controllers/api/ExecutionControllerTest.java
@@ -498,59 +498,4 @@ class ExecutionControllerTest {
 
         assertThat(error.getStatus(), is(HttpStatus.UNPROCESSABLE_ENTITY));
     }
-
-    @Test
-    void shouldRemoveLabelsFromExecution() throws QueueException, TimeoutException {
-
-        String flowId = "restart_with_inputs";
-        // Create an execution with labels
-        Label label1 = new Label("key1", "value1");
-        Label label2 = new Label("key2", "value2");
-
-        Execution execution = client.toBlocking().retrieve(
-            GET(
-                "/api/v1/executions/webhook/" + TESTS_FLOW_NS + "/webhook-dynamic-key/webhook-dynamic-key"
-            ),
-            Execution.class
-        );
-
-        // First add some labels to the execution
-        Execution executionWithLabels = client.toBlocking().retrieve(
-            HttpRequest
-                .POST("/api/v1/executions/" + execution.getId() + "/labels",
-                    List.of(label1, label2)),
-            Execution.class
-        );
-
-        // Verify labels were added
-        assertThat(executionWithLabels, notNullValue());
-        assertThat(executionWithLabels.getLabels(), hasSize(2));
-        assertThat(executionWithLabels.getLabels().getFirst(), is(new Label("key1", "value1")));
-        assertThat(executionWithLabels.getLabels().get(1), is(new Label("key2", "value2")));
-
-        // Now remove one label by sending a request with only one label
-        Execution executionWithOneLabel = client.toBlocking().retrieve(
-            HttpRequest
-                .POST("/api/v1/executions/" + execution.getId() + "/labels",
-                    List.of(label1)),
-            Execution.class
-        );
-
-        // Verify one label was removed
-        assertThat(executionWithOneLabel, notNullValue());
-        assertThat(executionWithOneLabel.getLabels(), hasSize(1));
-        assertThat(executionWithLabels.getLabels().getFirst(), is(new Label("key1", "value1")));
-
-        // Now remove all labels by sending an empty list
-        Execution executionWithNoLabels = client.toBlocking().retrieve(
-            HttpRequest
-                .POST("/api/v1/executions/" + execution.getId() + "/labels",
-                    Collections.emptyList()),
-            Execution.class
-        );
-
-        // Verify all labels were removed
-        assertThat(executionWithNoLabels, notNullValue());
-        assertThat(executionWithNoLabels.getLabels(), hasSize(0));
-    }
 }
